% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_class.R
\name{ep_class_links}
\alias{ep_class_links}
\title{Discover which classes are linked by which properties}
\usage{
ep_class_links(endpoint, ...)
}
\arguments{
\item{endpoint}{A \code{sparql_endpoint} object.}

\item{...}{Passed on to methods (commonly \code{limit}).}
}
\value{
A data frame with columns \code{from_class}, \code{property}, \code{to_class}, \code{n}.
}
\description{
Discover which classes are linked by which properties
}
