% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_class.R
\name{ep_classes}
\alias{ep_classes}
\title{Discover the classes used by an endpoint}
\usage{
ep_classes(endpoint, ...)
}
\arguments{
\item{endpoint}{A \code{sparql_endpoint} object.}

\item{...}{Passed on to methods (commonly \code{limit}).}
}
\value{
A data frame with columns \code{class} and \code{n}.
}
\description{
Discover the classes used by an endpoint
}
