% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_class.R
\name{ep_properties}
\alias{ep_properties}
\title{Discover the properties (predicates) used by an endpoint}
\usage{
ep_properties(endpoint, ...)
}
\arguments{
\item{endpoint}{A \code{sparql_endpoint} object.}

\item{...}{Passed on to methods (commonly \code{limit}).}
}
\value{
A data frame with columns \code{property} and \code{n}.
}
\description{
Discover the properties (predicates) used by an endpoint
}
