% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_generic.R
\name{extract_schema_generic}
\alias{extract_schema_generic}
\title{Extract a schema from generic RDF data.}
\usage{
extract_schema_generic(source)
}
\arguments{
\item{source}{A SPARQL endpoint URL or a parsed \code{rdflib} compatible file.}
}
\value{
A schema list, see \code{.empty_schema()}.
}
\description{
Extract a schema from generic RDF data.
}
