% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_owl.R
\name{extract_schema_owl}
\alias{extract_schema_owl}
\title{Extract a schema from an OWL ontology}
\usage{
extract_schema_owl(source)
}
\arguments{
\item{source}{A SPARQL endpoint URL or a parsed rdflib compatible file.}
}
\value{
A schema list, see empty_schema().
}
\description{
Extract a schema from an OWL ontology
}
