% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_class.R
\name{generic_endpoint}
\alias{generic_endpoint}
\title{Generic 'SPARQL' 1.1 backend}
\usage{
generic_endpoint(
  url = character(0),
  name = NA_character_,
  headers = list(),
  timeout = 60
)
}
\arguments{
\item{url, name, headers, timeout}{See \code{\link[=sparql_endpoint]{sparql_endpoint()}}.}
}
\description{
The fallback backend used for any endpoint that isn't registered
under a more specific class.
}
