% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_name}
\alias{local_name}
\title{Short display name for a URI}
\usage{
local_name(uri)
}
\arguments{
\item{uri}{Character vector of URIs.}
}
\value{
A character vector the same length as \code{uri}.
}
\description{
Returns the fragment or last path segment of a URI, falling back to the
full URI when no \code{/} or \verb{#} delimiter is present.
}
\examples{
local_name("http://xmlns.com/foaf/0.1/Person")
local_name("http://www.w3.org/2002/07/owl#Class")
}
