% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{namespace_of}
\alias{namespace_of}
\title{Namespace portion of a URI}
\usage{
namespace_of(uri)
}
\arguments{
\item{uri}{Character vector of URIs.}
}
\value{
A character vector the same length as \code{uri}.
}
\description{
Extracts the "namespace" portion of a URI, keeping whichever delimiter
(\code{/} or \verb{#}) was actually used, so it can be compared directly against
known vocabulary namespace strings.
}
\examples{
namespace_of("http://xmlns.com/foaf/0.1/Person")
namespace_of("http://www.w3.org/2002/07/owl#Class")
}
