% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_scheme}
\alias{normalize_scheme}
\title{Normalise a URI's scheme}
\usage{
normalize_scheme(x)
}
\arguments{
\item{x}{Character vector of URIs.}
}
\value{
A character vector the same length as \code{x}.
}
\description{
Rewrites \verb{https://} to \verb{http://} so namespace comparisons aren't
scheme-sensitive (many vocabularies are inconsistently referenced with
either scheme).
}
\examples{
normalize_scheme("https://schema.org/")
}
