<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{sparql_query}
\alias{sparql_query}
\title{Send 'SPARQL' query against an endpoint through HTTP request.}
=======
\name{sparql_query}
\alias{sparql_query}
\title{Send a SPARQL Query}
>>>>>>> 015165589fa525d44ce108c1da6996a2c333b790
\usage{
sparql_query(url, query, timeout = 60)
}
\arguments{
\item{url}{Character scalar. The SPARQL endpoint URL.}
<<<<<<< HEAD

\item{query}{Character scalar. A SPARQL query.}

\item{timeout}{Numeric scalar. Timeout in seconds.}
}
\value{
The parsed response in Json Format.
}
\description{
Send 'SPARQL' query against an endpoint through HTTP request.
}
=======
\item{query}{Character scalar. A SPARQL query.}
\item{timeout}{Numeric scalar. Timeout in seconds.}
}
\value{A data frame containing the parsed SPARQL response.}
\description{Send a SPARQL query against an endpoint through an HTTP request.}
>>>>>>> 015165589fa525d44ce108c1da6996a2c333b790
