% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{sparql_query}
\alias{sparql_query}
\title{Send 'SPARQL' query against an endpoint through HTTP request.}
\usage{
sparql_query(url, query, timeout = 60)
}
\arguments{
\item{url}{Character scalar. The SPARQL endpoint URL.}

\item{query}{Character scalar. A SPARQL query.}

\item{timeout}{Numeric scalar. Timeout in seconds.}
}
\value{
The parsed response in Json Format.
}
\description{
Send 'SPARQL' query against an endpoint through HTTP request.
}
